chore: move next.toKeep.models to models
