+++ server/src/common/model/effectifs.model/parts/effectif.field.locker.part.js
@@ -1,4 +1,4 @@
-import { object, boolean } from "../../../json-schema/jsonSchemaTypes.js";
+import { object, boolean } from "../../json-schema/jsonSchemaTypes.js";
 
 const adresseLocker = object({
   numero: boolean(),
