+++ server/src/common/model/fiabilisationUaiSiret.model.js
@@ -1,5 +1,5 @@
-import { FIABILISATION_TYPES } from "../../constants/fiabilisationConstants.js";
-import { object, date, objectId, string, stringOrNull } from "../json-schema/jsonSchemaTypes.js";
+import { FIABILISATION_TYPES } from "../constants/fiabilisationConstants.js";
+import { object, date, objectId, string, stringOrNull } from "./json-schema/jsonSchemaTypes.js";
 
 export const collectionName = "fiabilisationUaiSiret";
