+++ server/src/common/model/jwtSessions.model.js
@@ -1,4 +1,4 @@
-import { object, string, objectId } from "../json-schema/jsonSchemaTypes.js";
+import { object, string, objectId } from "./json-schema/jsonSchemaTypes.js";
 
 export const collectionName = "jwtSessions";
