+++ server/src/common/model/organismes.model.js
@@ -1,11 +1,11 @@
 import Joi from "joi";
-import { object, objectId, string, date, arrayOf, boolean, integer } from "../json-schema/jsonSchemaTypes.js";
-import { adresseSchema } from "../json-schema/adresseSchema.js";
-import { RESEAUX_CFAS } from "../../constants/networksConstants.js";
-import { NATURE_ORGANISME_DE_FORMATION } from "../../utils/validationsUtils/organisme-de-formation/nature.js";
-import { schemaValidation } from "../../utils/schemaUtils.js";
-import { siretSchema, uaiSchema } from "../../utils/validationUtils.js";
-import { FIABILISATION_TYPES } from "../../constants/fiabilisationConstants.js";
+import { object, objectId, string, date, arrayOf, boolean, integer } from "./json-schema/jsonSchemaTypes.js";
+import { adresseSchema } from "./json-schema/adresseSchema.js";
+import { RESEAUX_CFAS } from "../constants/networksConstants.js";
+import { NATURE_ORGANISME_DE_FORMATION } from "../utils/validationsUtils/organisme-de-formation/nature.js";
+import { schemaValidation } from "../utils/schemaUtils.js";
+import { siretSchema, uaiSchema } from "../utils/validationUtils.js";
+import { FIABILISATION_TYPES } from "../constants/fiabilisationConstants.js";
 
 export const collectionName = "organismes";
