+++ server/src/common/model/permissions.model.js
@@ -1,6 +1,6 @@
 import Joi from "joi";
-import { object, objectId, objectIdOrNull, string, date, arrayOf, boolean } from "../json-schema/jsonSchemaTypes.js";
-import { schemaValidation } from "../../utils/schemaUtils.js";
+import { object, objectId, objectIdOrNull, string, date, arrayOf, boolean } from "./json-schema/jsonSchemaTypes.js";
+import { schemaValidation } from "../utils/schemaUtils.js";
 
 export const collectionName = "permissions";
