+++ server/src/common/model/uploads.model/uploads.model.js
@@ -1,5 +1,5 @@
-import { preDefinedModels } from "../../../constants/models/upload.models.part.js";
-import { object, objectId, array, date, arrayOf, any, string, boolean } from "../../json-schema/jsonSchemaTypes.js";
+import { preDefinedModels } from "../../constants/models/upload.models.part.js";
+import { object, objectId, array, date, arrayOf, any, string, boolean } from "../json-schema/jsonSchemaTypes.js";
 import { documentSchema } from "./parts/document.part.js";
 
 export const collectionName = "uploads";
