+++ server/src/common/model/userEvents.model.js
@@ -1,4 +1,4 @@
-import { object, objectId, string, date, stringOrNull, any } from "../json-schema/jsonSchemaTypes.js";
+import { object, objectId, string, date, stringOrNull, any } from "./json-schema/jsonSchemaTypes.js";
 
 export const collectionName = "userEvents";
