+++ server/tests/integration/db/indexes/formations.indexes.test.js
@@ -1,7 +1,7 @@
 import { strict as assert } from "assert";
 import { createIndexes, dropIndexes } from "../../../../src/common/model/indexes/index.js";
 import { getDbCollectionIndexes } from "../../../../src/common/mongodb.js";
-import formationsModelDescriptor from "../../../../src/common/model/next.toKeep.models/formations.model.js";
+import formationsModelDescriptor from "../../../../src/common/model/formations.model.js";
 import { formationsDb } from "../../../../src/common/model/collections.js";
 
 describe("Formations Indexes", () => {
