+++ server/tests/integration/db/indexes/users.indexes.test.js
@@ -1,7 +1,7 @@
 import { strict as assert } from "assert";
 import { createIndexes, dropIndexes } from "../../../../src/common/model/indexes/index.js";
 import { getDbCollectionIndexes } from "../../../../src/common/mongodb.js";
-import usersModelDescriptor from "../../../../src/common/model/previous.models/users.model.js";
+import usersModelDescriptor from "../../../../src/common/model/users.model.js";
 import { createUser } from "../../../../src/common/actions/users.actions.js";
 
 describe("Users Indexes", () => {
